arm64: config: Enable config PLATFORM_TEGRA for Tegra210
Enable PLATFORM_TEGRA for tegra210 defconfigs. There is two flags
for Tegra architecture ARCH_TEGRA and PLATFORM_TEGRA. Tegra driver
depends on either or both in most of cases.

In order to avoid duplication of the configs and moving every Tegra
driver to depends on ARCH_TEGRA, add config PLATFORM_TEGRA as
preparation patch to stress the kernel.

On follow on, the PLATFORM_TEGRA will be removed from Tegra kernel
source.

bug 200256189

Change-Id: Ia417e54d6d4ae27429945114da3499fdfa383faf
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/1292021