platform: tegra: Move powergate ops to common includes
The powergate driver provides the common APIs for powergating to
client. The implementation of these APIs are different for different
SoCs. The SoC specific implementation implement the common ops which
is used by common powergate driver.

Currently, the structure of powergate ops are defined in multiple places.
Move this to common header so that all SoCs implementation can refer this
header instead of redefining for them.

bug 200257351

Change-Id: I3745e400686a23981c6f8e14be5316f0e29c42a5
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/1294566
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
Reviewed-by: Venkat Reddy Talla <vreddytalla@nvidia.com>
GVS: Gerrit_Virtual_Submit