v4l2: Increase event subscription queue size
Increase the size of event subscription queue to 2.

If the internal event queue for a particular subscribed event
becomes full, then the oldest event in that queue will be dropped.
Since we need both the on and off signals to our subscription,
it is good to increase the queue to 2.

Bug 1787461
Bug 1824769

Change-Id: Ic38517639448b10e7723d18ae91e1c6e793a0563
Signed-off-by: Gigon Bae <gbae@nvidia.com>
Reviewed-on: http://git-master/r/1294535
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
Reviewed-by: Hu He <hhe@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bhanu Murthy V <bmurthyv@nvidia.com>
Reviewed-by: Jihoon Bang <jbang@nvidia.com>