clk: tegra: use SHARED_LIMIT
Some 'shared clocks' exist to limit the 'clock bus' rate. When these clocks
are enabled, we don't have to set the rate to their requested rate. We merely
need to ensure the limits are adhered to. This is accomplished by setting the
TEGRA_SHARED_BUS_RATE_LIMIT flag. This is mostly important for 'floor' clocks
which enforce a certain minimum rate. 'ceiling' clocks already have a flag but
it doesn't hurt to set this one as well.

Bug  200267979

Change-Id: Iddaa02cd41826a5617c6107d726ed260dcec7867
Signed-off-by: Peter De Schrijver <pdeschrijver@nvidia.com>
Reviewed-on: http://git-master/r/1296616
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>