platform: tegra: powergate: Get t1xx powergate ops
The APIs implemented in the powergate-ops-t1xx.c and declared in
powergate-ops-t1xx.h is only used in the T210 powergate driver i.e.
powergate-t21x.c.

Hence move the content of this file to powergate-t21x.c and get rid
of extra file as well as SoCs specific config usage in the driver
file.

This helps in kernel unifications.

bug 200257351

Change-Id: I106631367eeecba785b11a02b13525a4c7afc03c
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/1295200
GVS: Gerrit_Virtual_Submit
Reviewed-by: Shardar Mohammed <smohammed@nvidia.com>