firmware: tegra: infinite wait for signs of life
On pre-silicon platforms, firmware is usually started after or in
parallell with CPU execution. Additionally, if bpmp has extra debug
traces enabled, firmware boot can take additional time. This induces a
race which can cause timeout when kernel is waiting for a sign of life
from bpmp. Instead of this, wait for bpmp signs-of-life indefinitely.
If bpmp is not present, it can be indicated by removing the bpmp DT
entry and we will not block kernel execution.

Change-Id: Id838ccb3b209e9002ad24bc562ad2686925348d3
Signed-off-by: Sivaram Nair <sivaramn@nvidia.com>
Reviewed-on: http://git-master/r/1297386
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Timo Alho <talho@nvidia.com>
GVS: Gerrit_Virtual_Submit