drivers: cpuidle: bail out if DT node is not compatible
Currently, the DT of the idle states will be parsed first whether it's
compatible or not. This could cause some warning messages if it wan not
designed for generic ARM CPU idle driver. So we fix it to check the
compatible string before the parsing procedures.

Bug 200256189

Change-Id: I04326fdb9bf7241727578da607035493513bdd3a
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: http://git-master/r/1301190
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Venkat Reddy Talla <vreddytalla@nvidia.com>
Tested-by: Venkat Reddy Talla <vreddytalla@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>