media:tegra:mipical: Increase timeout
nvcs tests has mipical timeout intermittently with OV5693.
This is due to ov5693 is in LP00 after power up.
In this case, we have to start sensor streaming first.
But by the time mipical is ready to run, there is
a delay causing us lose the first LP11 signal.
In order to catch the next LP11 (vblank),
increase timeout to 10ms.

bug 1867425
Jira CHWI-755

Change-Id: I59bcef95a212a0fafef5b65ff2bdb2fdb194df61
Reviewed-on: http://git-master/r/1309855
Reviewed-by: Wenjia Zhou <wenjiaz@nvidia.com>
Tested-by: Wenjia Zhou <wenjiaz@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Jaspreet Kaur <jkaur@nvidia.com>
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
Reviewed-by: Shubham Chandra <shubhamc@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Jihoon Bang <jbang@nvidia.com>