kernel: arm64: Add "model name" for 64bit tasks
Currently we have restriction to display processor model name under
cpuinfo structure for 32 bit tasks only and thus, ubuntu unity center
application or the /proc/cpuinfo does not display processor model name
for 64 bit tasks.

With this we are removing the restriction of displaying model name and
/proc/cpuinfo and ubuntu unity center can display the processor
information to the user for 64 bit tasks.

Bug 200284198

Change-Id: Ic9d652f21997c3720b11b7b001bb2868b4ac4a0d
Reviewed-on: http://git-master/r/1314364
Reviewed-by: Ninad Malwade <nmalwade@nvidia.com>
Tested-by: Ninad Malwade <nmalwade@nvidia.com>
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bibek Basu <bbasu@nvidia.com>