sync: Remove WARN_ON on releasing sync_pt with signaling
android_fence_release currently WARNs if the released sync_pt
is in the active_list, i.e., signaling has been enabled for
it, but the fence has not yet been signaled.

This is not a cause for concern if the sync_pt implementation
(e.g. nvhost) handles this case - in nvhost's case, nvhost
has an internal representation of sync framework sync_pts so
it does not care if the struct sync_pt disappears. The only
result will be that once the fence is triggered, the timeline
will still be signaled but without any results.

Bug 1876787

Change-Id: Ib466fe1b486436f84fbe38d3875404ab38fdfa6a
Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
Reviewed-on: http://git-master/r/1316225
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Deepak Nibade <dnibade@nvidia.com>