dma-buf: fix leak for devices with lazy unmapping disabled
Currently, if the lazy unmapping is disabled, due to the
extra reference, the dmabuf attachment does not get freed at
both dma_buf_detach and dma_buf_release. This causes leak.

When dma_buf lazy unmapping is disabled, we need not hold
extra reference on the attachment. This way, the attachment
is freed on the dma_buf_detach when references are dropped
and thus freed.

Bug 200282925

Change-Id: Ie4678cf53aa23480fac548b8dab3673dfff333e8
Signed-off-by: Sri Krishna chowdary <schowdary@nvidia.com>
Reviewed-on: http://git-master/r/1316976
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>