thermal: tegra: add support for GPU hw-throttle
Add support to trigger pulse skippers on the GPU when a HOT trip
point is triggered. The pulse skippers can be signalled to throttle
at lo, medium and high depths\levels.

Also allow the throttle depth to be configured through DT and
introduce a new DT binding for the same.

Bug 200233003

Change-Id: I775135aadfa1343688ae5253113025dd9ca56cf4
Signed-off-by: Srikar Srimath Tirumala <srikars@nvidia.com>
Reviewed-on: http://git-master/r/1458106
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Jon Mayo <jmayo@nvidia.com>