soc/tegra: sysedp: Fix registering of consumers via sysfs
SYSEDP consumers can be registered via the sysfs node
'/sys/power/system_edp/consumer_register' by writing a string with
the following format:

'consumer_name state0, state1, …, stateN'

Currently, this is not working because the function strcspn() which
parses the string that is written is only looking for the first
newline character and not newline and space characters. This is
causing the entire string to treated as the name, which might be
bigger than the max permitted, and there to be no states. Fix this
by searching the string for either the first space or newline.

Bug 1811732

Change-Id: I13832f7dbb02f013ecb3c1c06a0b094b3e0aa202
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: http://git-master/r/1326976
(cherry picked from commit cc45e8f58c49ed5806a8a1132b6522ec4cef52bf)
Reviewed-on: http://git-master/r/1456975
GVS: Gerrit_Virtual_Submit
Reviewed-by: Shreshtha Sahu <ssahu@nvidia.com>
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>