clk: tegra: bpmp: add "staged" clk provider
Add a "staged" clk provider for situtions where only a partial set of
clks are supported and we want to let the drivers use what is
avaialable. We implement this by first querying the bpmp firmware to see
if the asked clk is supported. If yes, we return that clk to the caller.
If not, then we create a dummy clk that use it.

JIRA BPMP-429

Change-Id: I5a1e4ec01e83d5e01cf2f18c1592f9f0e26a7290
Signed-off-by: Sivaram Nair <sivaramn@nvidia.com>
Reviewed-on: http://git-master/r/1456476
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
(cherry picked from commit d457f70cc459ac96a554a80e06486857aca3a09e)