ASoC: ad193x: Add support for setting the DAC DAI format
By default the ad193x DAC is configured for TDM mode and is not
reconfigured based upon the actual mode being used by the machine.
Update the ad913x driver to configure the DAC DAI format based upon
the actual mode being used.

It is odd that the ADC is configured for TDM mode for I2S and not the
'STEREO' mode, but as we are not using the ADC leave this as-is for
now.

Bug 200286098

Change-Id: Ie9e6039b841d57c4f059da95180694f82e5135b9
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: http://git-master/r/1457888
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bibek Basu <bbasu@nvidia.com>