net: wireless: bcmdhd: Remove ASSERT in dhd_wlfc.c
Issue: During SAP streaming - "!((PKTLINK(p) != NULL)
 && (PKTLINK(p) != p))": file "dhd_wlfc." prints are
seen in kernel logs.

The skb_segment() in "net/core/skbuff.c" is storing
the tail pointer in segs->prev.The same list is
passed to dhd driver which is causing assert.

Since network driver is optimizing the list, expecting
previous pointer to be NULL or head is not valid.
So we should remove the Assert which checks for this

Fix: Need to remove the ASSERT(!((PKTLINK(p) != NULL
) && (PKTLINK(p) != p))) in _dhd_wlfc_prec_enque()

bug 200307719

Change-Id: I7c5a539a452be41fd13fd9e24e1a24cf27412228
Signed-off-by: CH L V Sai Kiran <saik@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1523030
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Manikanta Maddireddy <mmaddireddy@nvidia.com>
Reviewed-by: Om Prakash Singh <omp@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>