usb: otg: sync device mode on at resuming
With xotg enabled, we can observe that 'device mode on' is
after xudc resume as it is managed by the otg fsm which uses
workqueue mechanism to do callbacks and a delay for calls is
expected.

However this may cause an issue if 'device mode on' is invoked too
late and happens to run into the next coming system suspend process,
then immediate wakeup due to this might be seen as follows:

[  129.739436] Resume caused by WAKE76, 3530000.xhci
[  129.739436] Resume caused by WAKE78, 3530000.xhci

This commit fixes this by forcely synchronizing devcie mode at
xudc resuming, the sequences will be the same as without enabling
otg.

Adding new function call in otg core for this purpose:
void usb_otg_flush_work(struct usb_otg *otg)

Bug 200331125

Change-Id: I97a6c64ac367d75efdb61e0de6ba3a8af5f7a332
Signed-off-by: Hans Yang <hansy@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1563088
GVS: Gerrit_Virtual_Submit
Reviewed-by: Henry Lin <henryl@nvidia.com>
Reviewed-by: Sing-Han Chen <singhanc@nvidia.com>
Reviewed-by: Ashutosh Jha <ajha@nvidia.com>