security:tlk: enable CONFIG_TRUSTED_LITTLE_KERNEL
This patch enables CONFIG_TRUSTED_LITTLE_KERNEL option in the android
defconfig because it is needed for android customers only.

This patch removed unnecessary functions from ote_comms.c file.

te_restore_keyslots which is a TOS specific function is renamed for
both OSes as {tlk|trusty}_restore_keyslots and the caller is updated
in a separate patch

Bug 200346933

Change-Id: Id6cc3f328a886cbe39e034ce2d875583c9ee56e6
Signed-off-by: Pranami Bhattacharya <pranamib@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1571782
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>