soc/tegra: gpu-edp: Put the GPU clock on error
The GPU clock is retrieved using the 'clk_get_sys()' API during the
probe of the GPU EDP driver and so must be put with 'clk_put()' if
the probe fails. Currently, 'clk_put()' is not being called for the
GPU clock on failure and so fix this.

Bug 1811732

Change-Id: Iea8a2e6d9010e55122ce2e62effeb34b6ae42ffa
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1577059
GVS: Gerrit_Virtual_Submit
Reviewed-by: Timo Alho <talho@nvidia.com>
(cherry picked from commit 9d10e9c1e4423324a398c5f01183038e1578cfb0)