soc/tegra: gpu-edp: Fix GPU V/F table
The GPU EDP driver specifies what function should be used to calculate
the GPU rail voltage for a specific GPU frequency. Currently, the GPU
EDP driver specifies the 'tegra_dvfs_predict_millivolts()' function
and by using this function the voltage returned is dependent on a
thermal scaling index defined by the DVFS driver and the voltage
calculation varies with temperature. The problem with this is that the
GPU EDP driver caches the V/F table and when the temperature changes
it is not updated. This may allow the GPU to operate at a frequency
which is too high for a given power threshold.

The GPU EDP driver should use the function,
'tegra_dvfs_predict_mv_at_hz_max_tfloor()', to calculate the voltage
to provide the worst case voltage based upon the maximum thermal floor
and hence, does not vary with temperature.

Bug 2003470

Change-Id: Iedad9cd93c87ed605b6df72f8badd62ccc3822bf
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1578681
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Timo Alho <talho@nvidia.com>