clk: tegra: Fix maximum audio sync clock for Tegra210
The maximum frequency supported for I2S on Tegra210 is 24.576MHz.
However, the maximum I2S frequency is limited to 24MHz because that
is the maximum frequency of the audio sync clock. Increase the
maximum audio sync clock frequency to 24.576MHz for Tegra210 in order
to support 24.576MHz for I2S.

Update the tegra_clk_register_sync_source() function so that it does
not set the initial rate for the sync clocks and use the clock init
tables to set the initial rate instead.

Please note that this is a regression from Linux v3.10.

Bug 2020021

Change-Id: If0a7b4d4f11acd86e9a7c4b7a077985ed6aef2ce
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1594562
Reviewed-by: Peter De Schrijver <pdeschrijver@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ravindra Lokhande <rlokhande@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>