soc: tegra: Update DMIC maximum frequency
When attempting to capture audio from DMIC3 on Jetson TX1 after boot
with sampling frequencies of 16kHz or greater the following errors
are seen when enabling the DMIC clock and causes the audio capture
to fail.

 tegra-dvfs: rate 12287995 too high for dvfs on dmic3
 tegra210-dmic tegra210-dmic.2: clk_enable failed: -22

The problem is that the Tegra210 DVFS driver is limiting the DMIC
frequency to 12.19MHz, although the DMIC supports a maximum frequency
of 12.288MHz (48kHz * 256 over-sampling). The 12.19MHz limit came from
the Tegra210 DVFS frequency spreadsheet that describes the maximum
frequencies for all interfaces/peripherals. This spreadsheet is being
updated to increase the maximum frequency for the DMIC to 12.288MHz.
Therefore, update the Tegra210 DVFS driver to align with this new
maximum frequency for the DMIC interfaces.

Bug 2029031

Change-Id: I44a49013e67b19ee6ba88c71dbf1f83fcff408bf
Reviewed-on: https://git-master.nvidia.com/r/1612395
Reviewed-by: Mohan Kumar D <mkumard@nvidia.com>
Reviewed-by: Jonathan Hunter <jonathanh@nvidia.com>
Tested-by: Jonathan Hunter <jonathanh@nvidia.com>
Reviewed-by: Peter De Schrijver <pdeschrijver@nvidia.com>
Reviewed-by: Aleksandr Frid <afrid@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Timo Alho <talho@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>