drivers: camera: Fix calibration for CSI2
CSI2 mipi lanes calibration is driven by sensor
state. As soon as the mipi pad is powered on the
sensor is also powered/streamed to reach the standby
state for performing the calibration during the
initialization sequence of CSI2 HW.
Address issues with loadable sensor modules, calibration
will be performed only when the sensor is bound.

Bug 200324656

Change-Id: I63821287514c1202570fc78a287a01882ff69fcc
Signed-off-by: David Bang <dbang@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1565678
Reviewed-on: https://git-master.nvidia.com/r/1620526
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bhanu Murthy V <bmurthyv@nvidia.com>
Reviewed-by: Jihoon Bang <jbang@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>