dmaengine: tegra210-adma: Add pause/resume support
The tegra-alt PCM driver specifies the flags SNDRV_PCM_INFO_PAUSE and
SNDRV_PCM_INFO_RESUME which indicates that the DMA driver supports
pausing and resuming of a DMA channel. However, the ADMA driver does
not populate the pause/resume function pointers for the ADMA driver
and so when the sound core attempts to pause a DMA channel, the
channel is not paused as expected.

Given that the ADMA controller supports pausing and resuming of a DMA
transfer, populate the DMAENGINE pause/resume function pointers for
the ADMA driver.

Please note that the 'wait_for_burst_complete' parameter that was
being passed to the tegra_adma_pause() function was not being used
within the function and so this parameter has been removed. This
aligns with both the upstream and Linux v4.9 kernels that don't have
such a parameter either.

Bug 200363507

Change-Id: I72d36bb00e094ca0b1f4043cad4e22ebf535650a
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1636470
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Mohan Kumar D <mkumard@nvidia.com>
Reviewed-by: Asha Talambedu <atalambedu@nvidia.com>
Reviewed-by: Sameer Pujar <spujar@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>