r8152: enable napi/netif after set_rx_mode
When a link-up condition is detected, make sure that set_rx_mode() is
called before informing napi/netif that the device is ready. Without this,
packets can be transmitted between the time that napi/netif are informed
the link is up and the time that set_rx_mode() is called. If the remote
host responds to those packets very quickly, the responses can also be
received before set_rx_mode() has run or completed. This sometimes causes
the received response packets to be completely missed (dropped by an RX
filter, or corrupted and dropped in HW), and so the kernel network stack
never gets to see them. In particular, this negatively affects systems
that use NFS for the root file-system during system suspend/resume; when
the system resumes, it sends out new requests to the NFS server, and the
first few responses are sometimes dropped due to the aforementioned issue,
which then requires TCP retransmission to recover, which significantly
extends system resume time.

Bug 2034208
Bug 1947539
Bug 1882292

Change-Id: Iea7c5322c848c48b9747ec3b70cb21cecbbdd243
Signed-off-by: Stephen Warren <swarren@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1646855
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>