ARM64: configs: t18x: Disable ADSP
By default L4T enables the ALSA restore service and so on boot the
ALSA configuration is restored. When this happens the following
warnings are observed ...

 [   16.848621] tegra210-adsp tegra210-adsp: ADSP is not booted yet
 [   16.848733] tegra210-adsp tegra210-adsp: ADSP is not booted yet
 [   16.848821] tegra210-adsp tegra210-adsp: ADSP is not booted yet
 [   16.848907] tegra210-adsp tegra210-adsp: ADSP is not booted yet
 [   16.848993] tegra210-adsp tegra210-adsp: ADSP is not booted yet
 [   16.849067] tegra210-adsp tegra210-adsp: ADSP is not booted yet
 [   16.849141] tegra210-adsp tegra210-adsp: ADSP is not booted yet
 [   16.849215] tegra210-adsp tegra210-adsp: ADSP is not booted yet
 ...

The tegra18_defconfig used by L4T enables the ADSP driver and by
enabling the ADSP driver, when the ADSP driver is probed, many mixer
controls are registered for the ADSP driver with the sound core.
However, none of these mixer controls can actually be used until the
ADSP is booted, and if anyone attempts to configure them before the
ADSP is booted, the above warning is seen.

For L4T the ADSP is never booted and so the above warnings are seen on
boot when the ALSA restore service tries to restore them. For L4T the
ADSP is not used and so disable the ADSP in the tegra18_defconfig.

Bug 200359126

Change-Id: Ie6f515ef01cc05bc6ef787d463b88b657cd1aa74
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1652141
(cherry picked from commit 2e5ed2924c082a5b162b8b2cfd73500ebd9aadd8)
Reviewed-on: https://git-master.nvidia.com/r/1668914
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>