kbuild: don't warn on -Wattribute-alias
Currently, -Wattribute-alias causes a variety of warnings that are
breaking the build. This is getting sorted out on LKML, but as far as I
can tell, no patch for it has yet been merged. For more context, see
https://lkml.org/lkml/2017/12/5/538.

In the meantime, disable the warning. We should drop this patch when a
proper fix has been merged into the kernel.

Signed-off-by: Martin Kelly <mkelly@xevo.com>
Signed-off-by: Matt Madison <matt@madison.systems>