fuse: tegra: disable mirroring during fuse burn
For T210, fuse mirroring is supported, making mirroring support flag
true/enabled. So that while burning the fuse, mirroring is disabled
and re-enabled after done. As if the mirroring is enabled during burn
then the data written to the fuse is not immediately visible without
a system reboot.

Bug 2310887

Change-Id: I7a1366643e9241311c19aa3e6d9529e8f6463851
Signed-off-by: kartik <kkartik@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1800841
GVS: Gerrit_Virtual_Submit
Reviewed-by: Shardar Mohammed <smohammed@nvidia.com>
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>